// src/proto-ts/build/make/tools/aconfig/aconfig_protos/protos/aconfig.ts
export enum flagState {
  ENABLED = 1,
  DISABLED = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum flagPermission {
  READ_ONLY = 1,
  READ_WRITE = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface flagMetadata {
  purpose?: flagMetadata_flagPurpose | undefined;
  storage?: flagMetadata_flagStorageBackend | undefined;
}

export enum flagMetadata_flagPurpose {
  PURPOSE_UNSPECIFIED = 0,
  PURPOSE_FEATURE = 1,
  PURPOSE_BUGFIX = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum flagMetadata_flagStorageBackend {
  /** UNSPECIFIED - unspecified, let aconfig choose one */
  UNSPECIFIED = 0,
  /** ACONFIGD - aconfigd based new storage */
  ACONFIGD = 1,
  /** DEVICE_CONFIG - device config based legacy storage */
  DEVICE_CONFIG = 2,
  /** NONE - no need for flag storage */
  NONE = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface tracepoint {
  /** path to declaration or value file relative to $TOP */
  source?: string | undefined;
  state?: flagState | undefined;
  permission?: flagPermission | undefined;
}

export interface parsedFlag {
  /**
   * Package to which the flag belongs (required)
   * See # package for format detail
   */
  package?:
    | string
    | undefined;
  /**
   * Name of the flag (required)
   * See # name for format detail
   */
  name?:
    | string
    | undefined;
  /**
   * Namespace the flag belongs to (required)
   * See # namespace for format detail
   */
  namespace?:
    | string
    | undefined;
  /** Textual description of the flag's purpose (required) */
  description?:
    | string
    | undefined;
  /** Single bug id related to the flag (required) */
  bug: string[];
  state?: flagState | undefined;
  permission?: flagPermission | undefined;
  trace: tracepoint[];
  /**
   * Indicates if the flag is permanently read-only and cannot be changed
   * via release configs (optional)
   * Default value false
   */
  isFixedReadOnly?:
    | boolean
    | undefined;
  /**
   * Indicates if the flag is exported and accessible beyond its originating container (optional)
   * Default value false
   */
  isExported?:
    | boolean
    | undefined;
  /**
   * Container the flag belongs to (optional)
   * See # container for format detail
   */
  container?:
    | string
    | undefined;
  /** Additional information about the flag, including its purpose and form factors (optional) */
  metadata?: flagMetadata | undefined;
}

export interface parsedFlags {
  parsedFlag: parsedFlag[];
}
// ...
export const flagMetadata: MessageFns<flagMetadata> = {
  encode(message: flagMetadata, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.purpose !== undefined && message.purpose !== 0) {
      writer.uint32(8).int32(message.purpose);
    }
    if (message.storage !== undefined && message.storage !== 0) {
      writer.uint32(16).int32(message.storage);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): flagMetadata {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseflagMetadata();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.purpose = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.storage = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): flagMetadata {
    return {
      purpose: isSet(object.purpose) ? flagMetadata_flagPurposeFromJSON(object.purpose) : 0,
      storage: isSet(object.storage) ? flagMetadata_flagStorageBackendFromJSON(object.storage) : 0,
    };
  },

  toJSON(message: flagMetadata): unknown {
    const obj: any = {};
    if (message.purpose !== undefined && message.purpose !== 0) {
      obj.purpose = flagMetadata_flagPurposeToJSON(message.purpose);
    }
    if (message.storage !== undefined && message.storage !== 0) {
      obj.storage = flagMetadata_flagStorageBackendToJSON(message.storage);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<flagMetadata>, I>>(base?: I): flagMetadata {
    return flagMetadata.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<flagMetadata>, I>>(object: I): flagMetadata {
    const message = createBaseflagMetadata();
    message.purpose = object.purpose ?? 0;
    message.storage = object.storage ?? 0;
    return message;
  },
};
// ...
export const tracepoint: MessageFns<tracepoint> = {
  encode(message: tracepoint, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.source !== undefined && message.source !== "") {
      writer.uint32(10).string(message.source);
    }
    if (message.state !== undefined && message.state !== 1) {
      writer.uint32(16).int32(message.state);
    }
    if (message.permission !== undefined && message.permission !== 1) {
      writer.uint32(24).int32(message.permission);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): tracepoint {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasetracepoint();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.source = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.state = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.permission = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): tracepoint {
    return {
      source: isSet(object.source) ? globalThis.String(object.source) : "",
      state: isSet(object.state) ? flagStateFromJSON(object.state) : 1,
      permission: isSet(object.permission) ? flagPermissionFromJSON(object.permission) : 1,
    };
  },

  toJSON(message: tracepoint): unknown {
    const obj: any = {};
    if (message.source !== undefined && message.source !== "") {
      obj.source = message.source;
    }
    if (message.state !== undefined && message.state !== 1) {
      obj.state = flagStateToJSON(message.state);
    }
    if (message.permission !== undefined && message.permission !== 1) {
      obj.permission = flagPermissionToJSON(message.permission);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<tracepoint>, I>>(base?: I): tracepoint {
    return tracepoint.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<tracepoint>, I>>(object: I): tracepoint {
    const message = createBasetracepoint();
    message.source = object.source ?? "";
    message.state = object.state ?? 1;
    message.permission = object.permission ?? 1;
    return message;
  },
};
// ...
export const parsedFlags: MessageFns<parsedFlags> = {
  encode(message: parsedFlags, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.parsedFlag) {
      parsedFlag.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): parsedFlags {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseparsedFlags();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.parsedFlag.push(parsedFlag.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): parsedFlags {
    return {
      parsedFlag: globalThis.Array.isArray(object?.parsedFlag)
        ? object.parsedFlag.map((e: any) => parsedFlag.fromJSON(e))
        : globalThis.Array.isArray(object?.parsed_flag)
        ? object.parsed_flag.map((e: any) => parsedFlag.fromJSON(e))
        : [],
    };
  },

  toJSON(message: parsedFlags): unknown {
    const obj: any = {};
    if (message.parsedFlag?.length) {
      obj.parsedFlag = message.parsedFlag.map((e) => parsedFlag.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<parsedFlags>, I>>(base?: I): parsedFlags {
    return parsedFlags.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<parsedFlags>, I>>(object: I): parsedFlags {
    const message = createBaseparsedFlags();
    message.parsedFlag = object.parsedFlag?.map((e) => parsedFlag.fromPartial(e)) || [];
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}